LopezCarlos - FixMinnorIssues

+++ Final_Java/src/entities/Prescription.java
@@ -57,7 +57,7 @@ public class Prescription {
 		return idItem;
 	}
 	
-	public int getTotal () {
+	public double getTotal () {
 		return total;
 	}
 }
